@jeanphilippeD 2019-11-13
Most fixes are now merged and the not checking for genesis info is more of an information than something we want to used.

Closing now.